fixed some conversion issues in Game::Start()

+++ Source/game.cpp
@@ -34,12 +34,12 @@ void Game::Start()
 	// creating walls 
 	auto window_width = (float)GetScreenWidth(); 
 	auto window_height = (float)GetScreenHeight(); 
-	float wall_distance = window_width / (wallCount + 1); 
+	float wall_distance = window_width / (float)(wallCount + 1); 
 	for (int i = 0; i < wallCount; i++)
 	{
 		Wall newWalls;
 		newWalls.position.y = window_height - 250; 
-		newWalls.position.x = wall_distance * (i + 1); 
+		newWalls.position.x = wall_distance * (float)(i + 1); 
 
 		Walls.push_back(newWalls);
